url registr process and long press bug fix

+++ app/src/main/java/com/puutaro/commandclick/util/file/FileSystems.kt
@@ -81,7 +81,10 @@ object FileSystems {
         fileName: String,
     ) {
         val filePath = File(dirPath, fileName)
-        if(!filePath.exists()) return
+        if(
+            !filePath.exists()
+            || !filePath.isFile
+        ) return
         try {
             filePath.delete()
         } catch (e: java.lang.Exception){
